// types.ts
export interface DiceBearOptions {
  seed?: string;
  flip?: boolean;
  backgroundColor?: string[];
  backgroundType?: string[];
  earrings?: string[];
  eyebrows?: string[];
  eyes?: string[];
  features?: string[];
  glasses?: string[];
  hair?: string[];
  hairColor?: string[];
  mouth?: string[];
  skinColor?: string[];
}